+++ dango/testing/src/lib.rs
@@ -2,6 +2,7 @@ mod account;
 pub mod constants;
 mod crypto;
 mod genesis;
+mod hyperlane;
 mod setup;
 
-pub use {account::*, crypto::*, genesis::*, setup::*};
+pub use {account::*, crypto::*, genesis::*, hyperlane::*, setup::*};
